+++ project-set/core/core-lib/src/main/java/com/rackspace/papi/service/context/container/ContainerConfigurationService.java
@@ -11,7 +11,7 @@ public interface ContainerConfigurationService {
    
    void setVia(String via);
    
-   int getContentBodyReadLimit();
+   Long getContentBodyReadLimit();
    
-   void setContentBodyReadLimit(int value);
+   void setContentBodyReadLimit(Long value);
 }
